requirements.txt
# AutoKR - 동영상 자막 자동 생성 시스템
# Python 3.9+ 필요

# 핵심 라이브러리
openai-whisper>=20231117  # Whisper STT
torch>=2.0.0              # PyTorch (GPU 가속)
torchaudio>=2.0.0         # 오디오 처리

# 오디오 처리
ffmpeg-python>=0.2.0      # FFmpeg Python 바인딩

# 번역
transformers>=4.35.0    # Hugging Face Transformers
sentencepiece>=0.1.99   # NLLB 토크나이저
protobuf>=3.20.0        # Protocol Buffers (transformers 의존성)

# 자막 생성
# SubtitleGenerator는 순수 Python으로 구현되어 추가 라이브러리 불필요

# 유틸리티
numpy>=1.24.0
